[refactor] Isolate JMX code, split config items.
