+++ extensions/modules/http-client/src/main/java/org/exist/xquery/modules/httpclient/jmx/HttpClientCacheMXBean.java
@@ -19,7 +19,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.exist.xquery.modules.httpclient;
+package org.exist.xquery.modules.httpclient.jmx;
 
 /**
  * JMX MXBean interface for monitoring the {@link HttpClientFactory} client cache.
